feat(core): extract browser-safe types entry (spec 001 task 1)
Extract pure Profile/AuthType/DeploymentType/AuthConfig into types.ts
(zero imports), re-exported from config.ts for back-compat. Add
index.browser.ts (types, redact, confluence-url) and conditional
package.json exports so browser bundlers resolve the Node-free entry
while the CLI keeps the full barrel. Rename ambient types.d.ts to
ambient.d.ts to avoid a basename collision with the new types.ts.
