+++ packages/core/src/ambient.d.ts
