+++ packages/core/src/index.browser.ts
@@ -1,0 +1,12 @@
+/**
+ * Browser-safe entry point for `@atlcli/core`.
+ *
+ * Resolved by bundlers targeting the browser (via the `browser` export
+ * condition in package.json). Every module re-exported here MUST build for
+ * `--target=browser` with zero `node:`/`bun:` specifiers in its transitive
+ * graph — enforced by the CI browser-build gate (spec 001 §6).
+ */
+
+export * from "./types.js";
+export * from "./redact.js";
+export * from "./confluence-url.js";
